requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap4==3.0.1
django-stdimage==5.3.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
mysql==0.0.3
mysqlclient==2.0.3
Pillow==8.2.0
pipenv==2020.8.13
psycopg2-binary==2.8.6
pycosat==0.6.3
PyMySQL==1.0.2
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
